requirements.txt
honeycomb-beeline>=2.5.0
lambda_decorators